// src/redux/types/AccountActionType.ts
export enum AccountActionType {
  CHANGE_COMMON_CONFIG = "CHANGE_COMMON_CONFIG",
  LOGIN = "LOGIN",
  LOGOUT = "LOGOUT",
  REFRESH_TOKEN = "REFRESH_TOKEN",
  GET_USER_INFO = "USER_INFO",
  EXTERNAL_LOGIN = "EXTERNAL_LOGIN",
  CHANGE_LANGUAGE = "CHANGE_LANGUAGE",
  CHANGE_LANGUAGE_WITH_LAUNCH = "CHANGE_LANGUAGE_WITH_LAUNCH",
  CHANGE_CURRENCY = "CHANGE_CURRENCY",
  CHANGE_CURRENCY_WITH_LAUNCH = "CHANGE_CURRENCY_WITH_LAUNCH",
  CHANGE_WEBSITE = "CHANGE_WEBSITE",
  CHANGE_LOADING = "CHANGE_LOADING",
  CHANGE_ANONYMOUS = "CHANGE_ANONYMOUS",
  MERGE_CART = "MERGE_CART",
  CHANGE_ORDER_IDS = "CHANGE_ORDER_IDS",
  CHANGE_PRODUCT_IDS = "CHANGE_PRODUCT_IDS",
  SET_SELECTED_ADDRESS_ID = "SET_SELECTED_ADDRESS_ID",
  CHECK_EXIST_CARD = "CHECK_EXIST_CARD",
  CHANGE_FAVORITE_PRODUCT = "CHANGE_FAVORITE_PRODUCT",
  CHANGE_BID_PRODUCT = "CHANGE_BID_PRODUCT",
  CHANGE_DEVICE_ID = "CHANGE_DEVICE_ID",
  GET_POST_OFFICE = "GET_POST_OFFICE",
}